backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.5.0
python-dotenv==1.0.1
firebase-admin==7.1.0

